+++ reflex-dom-core/ChangeLog.md
@@ -19,6 +19,10 @@
 
 * Fix bug in hydration when dealing with unexpected HTML ([#361](https://github.com/reflex-frp/reflex-dom/pull/361)).
 
+* Export attributes used for controlling hydration at the element level
+  * "data-ssr" is now available as `Reflex.Dom.Builder.Immediate.hydratableAttribute`
+  * "data-hydration-skip" is now available as `Reflex.Dom.Builder.Immediate.skipHydrationAttribute`
+
 ## 0.5.3
 
 * Deprecate a number of old inflexible widget helpers in `Reflex.Dom.Widget.Basic`: